Move generic parts of StandardRequestHandler into ControlRequestHandler.
